requirements.txt
ultralytics>=8.0.0
sahi>=0.11.14
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=9.0.0